+++ controller/mangas/create.js
@@ -1,6 +1,6 @@
 import Manga from '../../models/Manga.js'
 
-const controller = {
+const createManga = {
     create : async (req, res) => { //create es el modelo
         try {
           req.body.author_id = "63fe8112f09373806fd89fe5"
@@ -20,4 +20,4 @@ const controller = {
         }
       } 
 }
-export default controller // exporta el controlador
+export default createManga
