+++ models/Manga.js
@@ -7,13 +7,13 @@ let schema = new mongoose.Schema(
     title: { type: String, required: true },
     cover_photo: { type: String, required: true },
     description:{ type: String, required: true },
-    category_id:{ type: mongoose.Types.ObjectId,ref: 'manga',required: true },
+    category_id:{ type: mongoose.Types.ObjectId,ref: 'categories_id',required: true },
   },
   {
     timestamps: true,
   }
 );
 
-let Manga = mongoose.model("mangas", schema);
+let Manga = mongoose.model("manga", schema);
 
 export default Manga;
