+++ models/User.js
@@ -4,13 +4,13 @@ let schema = new mongoose.Schema({
     name: {type: String, required: true},
     mail: { type: String, required: true },
     password: { type: String, required: true },
-    confirm_password:{ type: String, required: true },
     is_online: { type: Boolean },
     is_admin: { type: Boolean },
     is_author: { type: Boolean },
     is_company: { type: Boolean },
     is_verified: { type: Boolean, required: true },
-    verify_code: { type: String, required: true }
+    verify_code: { type: String, required: true },
+    photo: { type: String, required: true},
 }, {
     timestamps: true
 })
