--- models/data/mangas.js
@@ -1,8 +1,0 @@
-import "dotenv/config.js";
-import "../../config/database.js";
-import Manga from "../Manga.js";
-
-
-let manga = [];
-
-Manga.insertMany(manga);
